requirements.txt
slack-bolt>=1.18.0
anthropic>=0.40.0
gspread>=6.0.0
google-auth>=2.28.0
flask>=3.0.0
python-dotenv>=1.0.0
gunicorn>=21.0.0
